fix: extract dogfood block monolith
