+++ examples/docs/dogfood-block-registry-key.ts
@@ -1,0 +1,3 @@
+export function dogfoodBlockKey(packageName: string, blockName: string): string {
+  return `${packageName}/${blockName}`;
+}
