+++ examples/docs/dogfood-block-registry.ts
@@ -1,0 +1,118 @@
+import { DOGFOOD_BLOCK_PACKAGE, s } from './dogfood-block-common.js';
+import type { DogfoodBlockDefinition, DogfoodBlockRole } from './dogfood-block-common.js';
+import { dogfoodBlockKey } from './dogfood-block-registry-key.js';
+import { isDogfoodBlockRegistryEntry } from './dogfood-block-registry-entry.js';
+import { DOGFOOD_BLOCK_REGISTRY_BRAND } from './dogfood-block-registry-types.js';
+import type {
+  DogfoodBlockRegistryBrandCarrier,
+  DogfoodBlockRegistryEntry,
+} from './dogfood-block-registry-types.js';
+import { normalizeRequiredText } from './dogfood-block-text.js';
+
+export class DogfoodBlockRegistry {
+  readonly [DOGFOOD_BLOCK_REGISTRY_BRAND]!: true;
+  readonly #entries: readonly DogfoodBlockRegistryEntry[];
+  readonly #entriesByBlockKey: ReadonlyMap<string, DogfoodBlockRegistryEntry>;
+  readonly #entriesBySurfaceId: ReadonlyMap<string, DogfoodBlockRegistryEntry>;
+
+  constructor(entries: readonly DogfoodBlockRegistryEntry[]) {
+    if (!Array.isArray(entries)) {
+      throw new Error('dogfood block registry: entries must be an array');
+    }
+
+    Object.defineProperty(this, DOGFOOD_BLOCK_REGISTRY_BRAND, { value: true });
+
+    const entriesByBlockKey = new Map<string, DogfoodBlockRegistryEntry>();
+    const entriesBySurfaceId = new Map<string, DogfoodBlockRegistryEntry>();
+    const normalizedEntries: DogfoodBlockRegistryEntry[] = [];
+
+    entries.forEach((entry, index) => {
+      if (!isDogfoodBlockRegistryEntry(entry)) {
+        throw new Error(
+          `dogfood block registry: entry at index ${s(index)} was not created by dogfoodBlockRegistryEntry()`,
+        );
+      }
+
+      const blockKey = dogfoodBlockKey(entry.packageName, entry.blockName);
+      if (entriesByBlockKey.has(blockKey)) {
+        throw new Error(`dogfood block registry: duplicate block ${blockKey}`);
+      }
+      if (entriesBySurfaceId.has(entry.surfaceId)) {
+        throw new Error(`dogfood block registry: duplicate surface id ${entry.surfaceId}`);
+      }
+
+      entriesByBlockKey.set(blockKey, entry);
+      entriesBySurfaceId.set(entry.surfaceId, entry);
+      normalizedEntries.push(entry);
+    });
+
+    this.#entries = Object.freeze(normalizedEntries);
+    this.#entriesByBlockKey = entriesByBlockKey;
+    this.#entriesBySurfaceId = entriesBySurfaceId;
+    Object.freeze(this);
+  }
+
+  entries(): readonly DogfoodBlockRegistryEntry[] {
+    return Object.freeze([...this.#entries]);
+  }
+
+  blocks(): readonly DogfoodBlockDefinition[] {
+    return Object.freeze(this.#entries.map((entry) => entry.block));
+  }
+
+  blockNames(): readonly string[] {
+    return Object.freeze(this.#entries.map((entry) => entry.blockName));
+  }
+
+  surfaceIds(): readonly string[] {
+    return Object.freeze(this.#entries.map((entry) => entry.surfaceId));
+  }
+
+  roles(): readonly DogfoodBlockRole[] {
+    const roles = new Set<DogfoodBlockRole>();
+    this.#entries.forEach((entry) => roles.add(entry.role));
+    return Object.freeze([...roles]);
+  }
+
+  forSurface(surfaceId: string): DogfoodBlockRegistryEntry | undefined {
+    return this.#entriesBySurfaceId.get(normalizeRequiredText({
+      scope: 'dogfood block registry',
+      field: 'surfaceId',
+      value: surfaceId,
+    }));
+  }
+
+  forBlock(blockName: string, packageName = DOGFOOD_BLOCK_PACKAGE): DogfoodBlockRegistryEntry | undefined {
+    return this.#entriesByBlockKey.get(dogfoodBlockKey(
+      normalizeRequiredText({
+        scope: 'dogfood block registry',
+        field: 'packageName',
+        value: packageName,
+      }),
+      normalizeRequiredText({
+        scope: 'dogfood block registry',
+        field: 'blockName',
+        value: blockName,
+      }),
+    ));
+  }
+
+  with(entry: DogfoodBlockRegistryEntry): DogfoodBlockRegistry {
+    return new DogfoodBlockRegistry([...this.#entries, entry]);
+  }
+}
+
+export function dogfoodBlockRegistry(
+  entries: readonly DogfoodBlockRegistryEntry[],
+): DogfoodBlockRegistry {
+  return new DogfoodBlockRegistry(entries);
+}
+
+export function isDogfoodBlockRegistry(value: unknown): value is DogfoodBlockRegistry {
+  return Boolean(
+    value
+      && typeof value === 'object'
+      && Object.prototype.hasOwnProperty.call(value, DOGFOOD_BLOCK_REGISTRY_BRAND)
+      && (value as DogfoodBlockRegistryBrandCarrier)[DOGFOOD_BLOCK_REGISTRY_BRAND] === true,
+  );
+}
