+++ examples/docs/dogfood-block-text.ts
@@ -1,0 +1,25 @@
+export function normalizeRequiredText(input: {
+  readonly scope: string;
+  readonly field: string;
+  readonly value: unknown;
+}): string {
+  if (typeof input.value !== 'string') {
+    throw new Error(`${input.scope}: ${input.field} must be a string`);
+  }
+
+  const value = input.value.trim();
+  if (value === '') {
+    throw new Error(`${input.scope}: ${input.field} is required`);
+  }
+
+  return value;
+}
+
+export function optionalTrimmedText(value: unknown): string | undefined {
+  if (value === undefined) return undefined;
+  return normalizeRequiredText({
+    scope: 'dogfood block registry entry',
+    field: 'description',
+    value,
+  });
+}
