+++ examples/docs/i18n-debt-baseline.ts
@@ -1,0 +1,47 @@
+import type { DogfoodI18nDebtBaseline } from './i18n-debt.js';
+
+export const DOGFOOD_I18N_DEBT_BASELINE: DogfoodI18nDebtBaseline = Object.freeze({
+  total: 2372,
+  bySurface: Object.freeze({
+    'docs-app': 153,
+    'capture-main': 9,
+    'counter-block-demo': 48,
+    'dogfood-block-command-palette': 36,
+    'dogfood-block-common': 14,
+    'dogfood-block-docs-surface-data': 30,
+    'dogfood-block-docs-surface-parser': 11,
+    'dogfood-block-docs-surface-render': 8,
+    'dogfood-block-docs-surface-schema': 15,
+    'dogfood-block-docs-surface': 22,
+    'dogfood-block-documentation-article': 35,
+    'dogfood-block-footer-hint': 28,
+    'dogfood-block-guide-inspector-render': 7,
+    'dogfood-block-guide-inspector': 30,
+    'dogfood-block-help-overlay': 33,
+    'dogfood-block-navigation-list-render': 6,
+    'dogfood-block-navigation-list': 33,
+    'dogfood-block-notification-center': 36,
+    'dogfood-block-perf-hud': 33,
+    'dogfood-block-preview': 37,
+    'dogfood-block-registry-entries': 27,
+    'dogfood-block-registry-entry': 6,
+    'dogfood-block-registry': 6,
+    'dogfood-block-schema-utils': 1,
+    'dogfood-block-search-panel-render': 7,
+    'dogfood-block-search-panel': 40,
+    'dogfood-block-settings-menu-render': 7,
+    'dogfood-block-settings-menu': 34,
+    'dogfood-block-text': 2,
+    'dogfood-block-title-screen': 34,
+    'dogfood-block-workbench-render': 7,
+    'dogfood-block-workbench': 45,
+    'component-stories': 1453,
+    coverage: 1,
+    'dogfood-locale': 7,
+    'i18n-dogfood-authoring': 1,
+    'i18n-missing-localization': 3,
+    'release-title': 33,
+    'storybook-app': 32,
+    'storybook-workstation': 3,
+  }),
+});
